Added tasks.py celery service
