+++ templates/biodata/add_detail.html
@@ -10,7 +10,7 @@
 
     <a href="{% url 'biodata:biodata' %}" class="btn btn-primary btn-md" role="button"> Preview Details</a>
 
-    <a href="{% url 'biodata:upload_detail' %}" class="btn btn-primary btn-md" role="button"> Upload Detail</a>
+    <a href="{% url 'biodata:do_upload' %}" class="btn btn-primary btn-md" role="button"> Upload Detail</a>
 
     <a href="/admin" class="btn btn-primary btn-md" role="button"> View History</a>
