+++ templates/biodata/edit_detail.html
@@ -10,7 +10,7 @@
 
     <a href="{% url 'biodata:biodata' %}" class="btn btn-primary btn-md" role="button"> Preview Details</a>
 
-    <a href="{% url 'biodata:upload' %}" target="_blank" class="btn btn-primary btn-md" role="button"> Upload Data</a>
+    <a href="{% url 'biodata:do_upload' %}" target="_blank" class="btn btn-primary btn-md" role="button"> Upload Data</a>
 
     <a href="/admin" target="_blank" class="btn btn-primary btn-md" role="button"> View History</a>
